pyproject.toml
[project]
name = "rems-for-organisations"
version = "2.0.0"
description = "Resources and Event Management System for small organisations and clubs. Bulk mailer, certificate generation and much more."
authors = [
    { name = "Krishnakanth Alagiri", email = "39209037+bearlike@users.noreply.github.com" },
    { name = "Mahalakshumi V", email = "reach@mahav.dev" },
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "email-to (>=0.1.0,<0.2.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "alembic (>=1.16,<2.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
djlint = "^1.36.4"
flake8 = "^7.2.0"

